move hook outside of mastra index
